Shifted away from POC of switch statement and to separate HTTP mapping
Rather than run with a switch statement bound to the micro error, it felt a bit more extensible to
pull out a simple convert method in the spirit of convertCode. Improvement thoughts: there are
issues around the 1:1 mapping of the http codes, ie:
[http-grpc-status-mapping](https://github.com/grpc/grpc/blob/master/doc/http-grpc-status-mapping.md).
There is also a large discrepancy in ready to go available codes within go-micro/errors. Would be
good to add to that for more specific usage (along the lines of what grpc already offers).